api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.121.2",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
